client/pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = ""
authors = ["leo <leopold@lchappuis.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
mss = "^9.0.2"
pillow = "^11.0.0"
bleak = "^0.22.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
